+++ packages/flowchart/jest.config.js
@@ -1,7 +1,6 @@
 const { BaseJestConfig } = require('../../config/jest');
 module.exports = {
   ...BaseJestConfig,
-  setupFilesAfterEnv: ['jest-extended', './jest.setup.js'],
   moduleNameMapper: {
     '^lodash-es$': 'lodash',
     '^.+\\.(css|less)$': 'identity-obj-proxy',