+++ packages/maps/jest.config.js
@@ -1,8 +1,2 @@
 const { BaseJestConfig } = require('../../config/jest');
-module.exports = {
-  ...BaseJestConfig,
-  moduleNameMapper: {
-    '^lodash-es$': 'lodash',
-    '^.+\\.(css|less)$': 'identity-obj-proxy',
-  },
-};
+module.exports = BaseJestConfig;