+++ packages/plots/jest.config.js
@@ -1,5 +1,2 @@
 const { BaseJestConfig } = require('../../config/jest');
-module.exports = {
-  ...BaseJestConfig,
-  setupFilesAfterEnv: ['jest-extended', './jest.setup.js'],
-};
+module.exports = BaseJestConfig;