+++ packages/plots/src/components/waterfall/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useImperativeHandle, forwardRef } from 'react';
 import { Waterfall as G2plotWaterfall, WaterfallOptions as G2plotConfig } from '@antv/g2plot';
 import useChart from '../../hooks/useChart';
-import { getChart } from '../../util';
+import { getChart } from '../../utils';
 import { BaseConfig } from '../../interface';
 import ErrorBoundary from '../../errorBoundary';
-import ChartLoading from '../../util/createLoading';
+import ChartLoading from '../../utils/createLoading';
 
 export interface WaterfallConfig extends Omit<G2plotConfig, 'tooltip'>, BaseConfig<G2plotConfig> {}