+++ packages/plots/tests/plots/bidirectional-bar.test.tsx
@@ -2,10 +2,10 @@
 import React, { useRef, createRef } from 'react';
 import { create } from 'react-test-renderer';
 import { renderHook } from '@testing-library/react-hooks/server';
-import ReactDOM from 'react-dom';
+import { render } from '../../src/utils';
 import { act } from 'react-dom/test-utils';
 import BidirectionalBar from '../../src/components/bidirectional-bar';
-import ChartLoading from '../../src/util/createLoading';
+import ChartLoading from '../../src/utils/createLoading';
 import ErrorBoundary from '../../src/errorBoundary';
 
 const refs = renderHook(() => useRef());
@@ -108,7 +108,7 @@ describe('BidirectionalBar render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<BidirectionalBar {...props} {...chartProps} />, container);
+      render(<BidirectionalBar {...props} {...chartProps} />, container);
     });
     expect(chartRef).not.toBeUndefined();
     const canvas = container.querySelector('canvas');
@@ -133,7 +133,7 @@ describe('BidirectionalBar render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<BidirectionalBar {...props} {...chartProps} />, container);
+      render(<BidirectionalBar {...props} {...chartProps} />, container);
     });
     expect(chartRef.current.chart.views[0].getData().length).toEqual(data.length);
   });
@@ -152,7 +152,7 @@ describe('BidirectionalBar render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<BidirectionalBar {...props} {...chartProps} ref={refs} />, container);
+      render(<BidirectionalBar {...props} {...chartProps} ref={refs} />, container);
     });
     expect(refs.current.getChart().chart.views[0].getData().length).toEqual(data.length);
   });