+++ packages/plots/tests/plots/chord.test.tsx
@@ -2,10 +2,10 @@
 import React, { useRef, createRef } from 'react';
 import { create } from 'react-test-renderer';
 import { renderHook } from '@testing-library/react-hooks/server';
-import ReactDOM from 'react-dom';
+import { render } from '../../src/utils';
 import { act } from 'react-dom/test-utils';
 import Chord from '../../src/components/chord';
-import ChartLoading from '../../src/util/createLoading';
+import ChartLoading from '../../src/utils/createLoading';
 import ErrorBoundary from '../../src/errorBoundary';
 
 const refs = renderHook(() => useRef());
@@ -113,7 +113,7 @@ describe('Chord render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Chord {...props} {...chartProps} />, container);
+      render(<Chord {...props} {...chartProps} />, container);
     });
     expect(chartRef).not.toBeUndefined();
     const canvas = container.querySelector('canvas');
@@ -139,7 +139,7 @@ describe('Chord render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Chord {...props} {...chartProps} />, container);
+      render(<Chord {...props} {...chartProps} />, container);
     });
     expect(chartRef.current.chart.views[0].getData().length).toBe(data.length);
   });
@@ -159,7 +159,7 @@ describe('Chord render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Chord {...props} {...chartProps} ref={refs} />, container);
+      render(<Chord {...props} {...chartProps} ref={refs} />, container);
     });
     expect(refs.current.getChart().chart.views[0].getData().length).toBe(data.length);
   });