+++ packages/plots/tests/plots/dual-axes.test.tsx
@@ -2,10 +2,10 @@
 import React, { useRef, createRef } from 'react';
 import { create } from 'react-test-renderer';
 import { renderHook } from '@testing-library/react-hooks/server';
-import ReactDOM from 'react-dom';
+import { render } from '../../src/utils';
 import { act } from 'react-dom/test-utils';
 import DualAxes from '../../src/components/dual-axes';
-import ChartLoading from '../../src/util/createLoading';
+import ChartLoading from '../../src/utils/createLoading';
 import ErrorBoundary from '../../src/errorBoundary';
 
 const refs = renderHook(() => useRef());
@@ -126,7 +126,7 @@ describe('DualAxes render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<DualAxes {...props} {...chartProps} />, container);
+      render(<DualAxes {...props} {...chartProps} />, container);
     });
     expect(chartRef).not.toBeUndefined();
     const canvas = container.querySelector('canvas');
@@ -158,7 +158,7 @@ describe('DualAxes render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<DualAxes {...props} {...chartProps} />, container);
+      render(<DualAxes {...props} {...chartProps} />, container);
     });
     expect(chartRef.current.chart.views[0].getData()).toEqual([
       {
@@ -190,7 +190,7 @@ describe('DualAxes render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<DualAxes {...props} {...chartProps} ref={refs} />, container);
+      render(<DualAxes {...props} {...chartProps} ref={refs} />, container);
     });
     expect(refs.current.getChart().chart.views[0].getData()).toEqual([
       {