+++ packages/plots/tests/plots/progress.test.tsx
@@ -2,10 +2,10 @@
 import React, { useRef, createRef } from 'react';
 import { create } from 'react-test-renderer';
 import { renderHook } from '@testing-library/react-hooks/server';
-import ReactDOM from 'react-dom';
+import { render } from '../../src/utils';
 import { act } from 'react-dom/test-utils';
 import Progress from '../../src/components/progress';
-import ChartLoading from '../../src/util/createLoading';
+import ChartLoading from '../../src/utils/createLoading';
 import ErrorBoundary from '../../src/errorBoundary';
 
 const refs = renderHook(() => useRef());
@@ -92,7 +92,7 @@ describe('Progress render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Progress {...props} {...chartProps} />, container);
+      render(<Progress {...props} {...chartProps} />, container);
     });
     expect(chartRef).not.toBeUndefined();
     const canvas = container.querySelector('canvas');
@@ -118,7 +118,7 @@ describe('Progress render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Progress {...props} {...chartProps} />, container);
+      render(<Progress {...props} {...chartProps} />, container);
     });
     expect(chartRef.current.chart.getData()).toEqual([
       { type: 'current', percent: 0.25 },
@@ -138,7 +138,7 @@ describe('Progress render', () => {
       height: 160,
     };
     act(() => {
-      ReactDOM.render(<Progress {...props} {...chartProps} ref={refs} />, container);
+      render(<Progress {...props} {...chartProps} ref={refs} />, container);
     });
     expect(refs.current.getChart().chart.getData()).toEqual([
       { type: 'current', percent: 0.25 },