+++ packages/plots/CHANGELOG.md
@@ -1,4 +1,4 @@
-- 🐞 [修复数据更新时legend留存](https://github.com/ant-design/ant-design-charts/issues/620)
+- 🐞 [修复数据更新时 legend 留存](https://github.com/ant-design/ant-design-charts/issues/620)
 - 🐞 修复自定义 errorTemplate 非 callback 时无效
 
 ## 1.2.2
@@ -6,7 +6,6 @@
 `2022-09-07`
 
 - 🐞 [修复 Tooltip 部分配置类型错误](https://github.com/ant-design/ant-design-charts/issues/1555)
-
 
 ## 1.2.1
 
@@ -20,4 +19,3 @@
 
 - 🔥 [React18 兼容](https://github.com/ant-design/ant-design-charts/issues/1452)
 - `ContainerConfig` renamed `BaseConfig`
-